pyproject.toml
[project]
name = "fastapi-todo"
version = "0.1.0"
description = ""
authors = [
    {name = "JeissonRuiz02",email = "jeissonruizdev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "passlib[argon2] (>=1.7.4,<2.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]

[tool.poetry]
packages = [{include = "fastapi_todo", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
